types: Moved Degrees from core to types
