+++ core/src/display_object.rs
@@ -10,12 +10,11 @@ use crate::player::NEWEST_PLAYER_VERSION;
 use crate::prelude::*;
 use crate::tag_utils::SwfMovie;
 use crate::transform::Transform;
-use crate::types::Degrees;
 use crate::vminterface::{AvmType, Instantiator};
 use bitflags::bitflags;
 use gc_arena::{Collect, MutationContext};
 use ruffle_macros::enum_trait_object;
-use ruffle_types::numbers::Percent;
+use ruffle_types::numbers::{Degrees, Percent};
 use ruffle_types::string::{AvmString, WString};
 use std::cell::{Ref, RefMut};
 use std::fmt::Debug;
