+++ core/src/lib.rs
@@ -39,7 +39,6 @@ pub mod shape_utils;
 pub mod tag_utils;
 pub mod timer;
 mod transform;
-mod types;
 pub mod vminterface;
 mod xml;
